requirements.txt
# --- Data ingestion ---
finnhub-python==2.4.20
websocket-client==1.7.0
kafka-python==2.0.2

# --- Storage ---
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0

# --- Stream processing & detection ---
pyspark==3.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# --- Dashboard ---
streamlit>=1.32.0
plotly>=5.18.0

# --- Testing ---
pytest>=7.4.0
